Split up Key interface into PublicKey and PrivateKey interfaces
This makes it clearer whether a key is a public or private key, and uses
type safety to check that the right kind of key is being used.

We still share the same underlying structure type to store public and
private key data. But users will get a PublicKey or PrivateKey interface
type instead of the structure.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>